Splits Fn and Shift on {0, ..., 9, -, =} for QWERTY
This commit is intended to provide access to the F1-F12 keys on the F(x)tec Pro1.
I have not tested it.
When using the phone with a userland meant for Linux (either directly on the phone, or through remote control), it is common to find applications that expect F1-F12 keys to be available.
This commit would turn the redundant Fn+{1, ..., 9, 0, -, =} combination, which were hitherto aliased to Shift+{1, ..., 9, 0, -, =}, into {F1, ..., F12} when using the QWERTY layout.
Such a change would be more complicated to bring on the QWERTZ layout, as some of these combinations are already meaningful in that configuration, so it remains untouched.

I have chosen the {F1, ..., F12} keys because those are what I would like to map these combinations to, but I think ultimately the choice ought to be made by the user (and thus, in userspace). To my understanding, this cannot be done as long as the driver itself makes the aliasing. Of course, I would be remiss not to mention the best argument in favor of this change: as it stands, the F(x)tec phone does not have the F(x) tech. 😜